# back/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgendaVest API",
    "version": "2.0.0",
    "description": "Documentação oficial da API AgendaVest"
  },
  "servers": [
    {
      "url": "http://10.111.9.22:3001",
      "description": "Servidor AgendaVest"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "usuarioAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido no login do usuário"
      },
      "admAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido no login do administrador"
      }
    }
  },
  "paths": {
...
    "/fuvest": {
      "post": {
        "tags": [
          "Scraping"
        ],
        "summary": "Executar scraping da FUVEST",
        "responses": {
          "200": {
            "description": "Dados da FUVEST processados com sucesso"
          },
          "500": {
            "description": "Erro ao buscar ou salvar os dados da FUVEST"
          }
        }
      }
    },
    "/unicamp": {
      "post": {
        "tags": [
          "Scraping"
        ],
        "summary": "Executar scraping da UNICAMP",
        "responses": {
          "200": {
            "description": "Dados da UNICAMP processados com sucesso"
          },
          "500": {
            "description": "Erro ao buscar ou salvar os dados da UNICAMP"
          }
        }
      }
    },
    "/fuvest/provas": {
      "post": {
        "tags": [
          "Scraping"
        ],
        "summary": "Executar scraping das provas da FUVEST",
        "responses": {
          "200": {
            "description": "Provas da FUVEST processadas com sucesso"
          },
          "404": {
            "description": "FUVEST não encontrada no banco"
          },
          "500": {
            "description": "Erro ao buscar ou salvar as provas da FUVEST"
          }
        }
      }
    },
    "/unicamp/provas": {
      "post": {
        "tags": [
          "Scraping"
        ],
        "summary": "Executar scraping das provas da UNICAMP",
        "responses": {
          "200": {
            "description": "Provas da UNICAMP processadas com sucesso"
          },
          "404": {
            "description": "UNICAMP não encontrada no banco"
          },
          "500": {
            "description": "Erro ao buscar ou salvar as provas da UNICAMP"
          }
        }
      }
    }
  }
}